requirements.txt
flask==2.3.3
flask-cors==4.0.0
mysql-connector-python==8.2.0
python-dotenv==1.0.0
sentence-transformers>=2.7.0
chromadb>=0.5.0
Pillow>=10.0.0
